requirements.txt
# ── Core Dashboard ──────────────────────────────────────────────
dash==2.17.1
dash-bootstrap-components==1.6.0
plotly==5.22.0

# ── Database ─────────────────────────────────────────────────────
sqlalchemy==2.0.30

# ── Configuration ────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Data Sources ─────────────────────────────────────────────────
# FRED (Federal Reserve)
fredapi==0.5.1

# Yahoo Finance — mercados, índices, divisas, materias primas
yfinance==0.2.40

# World Bank
wbdata==1.0.0

# Criptomonedas
pycoingecko==3.1.0

# Peticiones HTTP genéricas (ECB, Eurostat, GDELT, NewsAPI)
requests==2.32.3
httpx==0.27.0

# ── Data Processing ──────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── Scheduler ────────────────────────────────────────────────────
apscheduler==3.10.4

# ── AI Engine (Módulo 15) ────────────────────────────────────────
anthropic==0.28.0

# ── Utilities ────────────────────────────────────────────────────
python-dateutil==2.9.0
pytz==2024.1
